requirements.txt
pydantic>=2.0.0
jinja2>=3.0.0
PyYAML>=6.0.0
click>=8.0.0
pathlib>=1.0.0
typing-extensions>=4.0.0